move GlyphSubstitutionTable to its original package
