+++ fontbox/src/main/java/org/apache/fontbox/ttf/SubstitutingCmapLookup.java
@@ -18,8 +18,6 @@ package org.apache.fontbox.ttf;
 
 import java.util.List;
 
-import org.apache.fontbox.ttf.gsub.GlyphSubstitutionTable;
-
 /**
  * A cmap lookup that performs substitution via the 'GSUB' table.
  *
